# RFID tracking web MVC/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RFID tracking API",
    "version": "1.0"
  },
  "paths": {
    "/api/Directors": {
      "get": {
        "tags": [
          "Directors"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Director"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Director"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Director"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Directors"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Director"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Director"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Director"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Director"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Director"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Director"
                }
              }
            }
          }
        }
      }
    },
    "/api/Directors/{id}": {
      "get": {
        "tags": [
          "Directors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Director"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Director"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Director"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Directors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Director"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Director"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Director"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Directors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Loans": {
      "get": {
        "tags": [
          "Loans"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loan"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loan"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loan"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Loans"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Loan"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Loan"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Loan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              }
            }
          }
        }
      }
    },
    "/api/Loans/{id}": {
      "get": {
        "tags": [
          "Loans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Loans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Loan"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Loan"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Loan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Loans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Police": {
      "get": {
        "tags": [
          "Police"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Police"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Police"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Police"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Police"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Police"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Police"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Police"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Police"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Police"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Police"
                }
              }
            }
          }
        }
      }
    },
    "/api/Police/{id}": {
      "get": {
        "tags": [
          "Police"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Police"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Police"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Police"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Police"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Police"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Police"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Police"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Police"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Director": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "mailAddress": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "format": "tel",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Loan": {
        "required": [
          "loanStart",
          "shooterId",
          "weaponId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "shooterId": {
            "type": "string",
            "format": "uuid"
          },
          "weaponId": {
            "type": "string",
            "format": "uuid"
          },
          "loanStart": {
            "type": "string",
            "format": "date-time"
          },
          "loanEnd": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Police": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "mailAddress": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "format": "tel",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}